separated most of the functions that used to be in main.c into other .c files to better follow OO practices
