+++ Lab2/SRC/buttons.h
@@ -1,0 +1,14 @@
+/*
+ * buttons.h
+ *
+ *  Created on: Feb 5, 2017
+ *      Author: nbeeten
+ */
+
+#ifndef SRC_BUTTONS_H_
+#define SRC_BUTTONS_H_
+
+void initButtons();
+void checkButtons();
+
+#endif /* SRC_BUTTONS_H_ */
