+++ Lab2/SRC/main.c
@@ -135,129 +135,12 @@ int main(){
 		break; //end of case ARM_POSITION
 	}
 
-}
-
-
+} //end of main
 
 
 
 int returnBITS(){
 	return ADMUX;
-}
-
-
-
-/**
- * @brief prints the time stamp, pot value, pot angle, pot milivolts
- */
-void printPotVal(){
-	double potAngle = 0;
-	double potmV = 0;
-
-	//save all values to things that can be used in a print function
-	potAngle = ADCtoAngle(ADCvalue);
-	potmV = ADCtoMiliV(ADCvalue);
-	timeVal = timerCnt * 0.5;
-
-	printf("%f, %d, %g, %f\n\r", timeVal, ADCvalue, potAngle, potmV);
-}
-
-void printPWMVal(){
-	double dutyCyc = 0;
-	//int freq = 0;
-	//int state = 0;
-
-	printf("%f, %d, %d, %d\n\r", dutyCyc, button, Thigh, ADCvalue);
-}
-/**
- * @brief sets the button global variable to be the button that was pressed.
- *
- * Checks starting at port 7 and works down
- */
-void checkButtons(){
-	if (PINC & 256){
-		button = 7;
-
-	}
-	else if(PINC & 128){
-		button = 6;
-
-	}
-	else if(PINC & 64){
-		button = 5;
-
-	}
-	else if(PINC & 16){
-		button = 4;
-	}
-	else{
-		button = 10;
-	}
-}
-
-/**
- * @brief inits the buttons on PORTB by setting all of PORTB pins to input
- */
-void initButtons(){
-	//sets all of PortC to be inputs
-	DDRC &= 0b00000000;
-}
-
-void initPWMPin(){
-	//sets all of PortB to be outputs;
-	DDRB &= 0b11111111;
-	PORTB &= 0b00000000;
-	PINB &= 0b00000000;
-
-}
-
-void generatePWM(unsigned int countTo){
-	//if timer reaches countTo
-	switch(output){
-
-	case 1:
-		if(PWMTimerCnt >= countTo){
-			//switch port
-			output = 0;
-			PWMTimerCnt = 0;
-			putCharDebug('p');
-			PORTB = 0b00000000;
-		}
-		break; //end case 1
-
-	case 0:
-		if(PWMTimerCnt >= countTo){
-			//switch port
-			output = 1;
-			PWMTimerCnt = 0;
-			putCharDebug('s');
-			PORTB = 0b00000010;
-		}
-		break; //end case 0
-	}
-}
-
-void outputPWM(){
-	switch(button){
-
-	case 7:
-		//generates a 100Hz signal
-		Thigh = 39;
-		generatePWM(Thigh);
-		break;
-
-	case 6:
-		//generates 20Hz signal
-		Thigh = 195;
-		generatePWM(Thigh);
-		break;
-
-	case 5:
-		//generates a 1Hz signal
-		Thigh = 3906;
-		generatePWM(Thigh);
-		break;
-	}
 }
 
 void ramp(){
