+++ Lab2/SRC/main.h
@@ -14,6 +14,9 @@
 #include "ADCcalc.h"
 #include "currentSense.h"
 #include "PIDmotors.h"
+#include "printStatements.h"
+#include "buttons.h"
+#include "pwm.h"
 
 #define MAX_ADC 1023
 #define ADC_CHANNEL 4
@@ -21,18 +24,8 @@
 #define DBUS1_CHANNEL 3
 
 
-
-double ADCtoMiliV(unsigned int potVal);
-void printPotVal();
-void printPWMVal();
 int returnBITS();
-void checkButtons();
-void initButtons();
-void initPWMPin();
-void generatePWM(unsigned int countTo);
-void outputPWM();
 void ramp();
-void PIDCntrl();
 
 
 #endif /* SRC_MAIN_H_ */
