+++ Lab2/SRC/printStatements.h
@@ -1,0 +1,14 @@
+/*
+ * printStatements.h
+ *
+ *  Created on: Feb 5, 2017
+ *      Author: nbeeten
+ */
+
+#ifndef SRC_PRINTSTATEMENTS_H_
+#define SRC_PRINTSTATEMENTS_H_
+
+void printPotVal();
+void printPWMVal();
+
+#endif /* SRC_PRINTSTATEMENTS_H_ */
